roles & permissions

+++ app/Services/User/UserService.php
@@ -2,8 +2,11 @@
 
 namespace App\Services\User;
 
+use App\Http\Requests\Admin\User\UserCreateRequest;
+use App\Http\Requests\Admin\User\UserUpdateRequest;
 use App\Models\User\User;
 use App\UseCases\Files\ImageService;
+use Auth;
 use Illuminate\Support\Facades\DB;
 use Storage;
 
@@ -16,28 +19,45 @@ class UserService
         $this->image = $image;
     }
 
-    public function update($data, User $user)
+    public function update(UserUpdateRequest $request, User $user)
     {
-        DB::transaction( function () use ($data, $user){
+        $data = collect($request->validated());
+
+//        dd($data);
+        DB::transaction(function () use ($data, $user) {
 
             $user->roles()->sync( $data->pull('roles') );
             //avatar
-            if( $data->has('avatar') ){
-                $data->put('avatar', $this->saveImage( $data->pull('avatar'), $user->id ));
+            if ($data->has('avatar')) {
+                $data->put('avatar', $this->saveImage($data->pull('avatar'), $user->id));
+            }
+            // activate
+            if ($data->has('active')) {
+                $data->put('email_verified_at', $data->get('active') ? now() : null);
             }
             //password
-            if( $pass = $data->pull('password') )
+            if ($pass = $data->pull('password')) {
                 $data->put('password', bcrypt($pass));
+            }
 
-            $user->update( $data->toArray() );
+            $user->update($data->toArray());
         });
     }
 
     /**
      * @param Post $post
      */
-    public function delete(User $user)
+    public function delete(User $user): void
     {
+        // нельзя удалить админа
+        if ($user->isAdmin()) {
+            throw new \RuntimeException('Admin user can not be deleted. Please change role.');
+        }
+        // нельзя удалить себя
+        if (Auth::id() === $user->id) {
+            throw new \RuntimeException('You can not delete yourself.');
+        }
+
         $this->removeImage($user);
         $user->delete();
     }
@@ -59,5 +79,9 @@ class UserService
                 ->upload(false)['user'];
     }
 
+    public function store(UserCreateRequest $request)
+    {
+        return User::create($request->all());
+    }
 
 }
